Tidy and organise userprint package
