+++ cf/actors/userprint/ui.go
@@ -1,0 +1,73 @@
+package userprint
+
+import (
+	"fmt"
+
+	. "github.com/cloudfoundry/cli/cf/i18n"
+	"github.com/cloudfoundry/cli/cf/models"
+	"github.com/cloudfoundry/cli/cf/terminal"
+)
+
+type SpaceUsersUiPrinter struct {
+	Ui               terminal.UI
+	UserLister       func(spaceGuid string, role string) ([]models.UserFields, error)
+	Roles            []string
+	RoleDisplayNames map[string]string
+}
+
+type OrgUsersUiPrinter struct {
+	Roles            []string
+	RoleDisplayNames map[string]string
+	UserLister       func(orgGuid string, role string) ([]models.UserFields, error)
+	Ui               terminal.UI
+}
+
+func (p *OrgUsersUiPrinter) PrintUsers(guid string, username string) {
+	for _, role := range p.Roles {
+		displayName := p.RoleDisplayNames[role]
+		users, err := p.UserLister(guid, role)
+		if err != nil {
+			p.Ui.Failed(T("Failed fetching org-users for role {{.OrgRoleToDisplayName}}.\n{{.Error}}",
+				map[string]interface{}{
+					"Error":                err.Error(),
+					"OrgRoleToDisplayName": displayName,
+				}))
+			return
+		}
+		p.Ui.Say("")
+		p.Ui.Say("%s", terminal.HeaderColor(displayName))
+
+		if len(users) == 0 {
+			p.Ui.Say(fmt.Sprintf("  "+T("No %s found"), displayName))
+		} else {
+			for _, user := range users {
+				p.Ui.Say("  %s", user.Username)
+			}
+		}
+	}
+}
+
+func (p *SpaceUsersUiPrinter) PrintUsers(guid string, username string) {
+	for _, role := range p.Roles {
+		displayName := p.RoleDisplayNames[role]
+		users, err := p.UserLister(guid, role)
+		if err != nil {
+			p.Ui.Failed(T("Failed fetching space-users for role {{.SpaceRoleToDisplayName}}.\n{{.Error}}",
+				map[string]interface{}{
+					"Error":                  err.Error(),
+					"SpaceRoleToDisplayName": displayName,
+				}))
+			return
+		}
+		p.Ui.Say("")
+		p.Ui.Say("%s", terminal.HeaderColor(displayName))
+
+		if len(users) == 0 {
+			p.Ui.Say(fmt.Sprintf("  "+T("No %s found"), displayName))
+		} else {
+			for _, user := range users {
+				p.Ui.Say("  %s", user.Username)
+			}
+		}
+	}
+}
